Remove maven_tools.dart
Move MavenDependency class definition to gradle_tools.dart
Removed unused BuildSummarizer pom.xml
Updated jnigen/CHANGELOG.md
Updated jnigen/README.md
Bump jnigen version.
